# update to AbstractPermutations-0.3