--- a/Project.toml
+++ b/Project.toml
@@ -10,8 +10,8 @@ PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
 [compat]
-AbstractPermutations = "0.2"
-GroupsCore = "0.4"
+AbstractPermutations = "0.3"
+GroupsCore = "0.5"
 PrecompileTools = "1"
 julia = "1.6"