--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "PermutationGroups"
 uuid = "8bc5a954-2dfc-11e9-10e6-cd969bffa420"
 authors = ["Marek Kaluba <kalmar@amu.edu.pl>", "tweisser <tillmann.weisser@web.de>"]
-version = "0.5.0"
+version = "0.6.0"
 
 [deps]
 AbstractPermutations = "36d08e8a-54dd-435f-8c9e-38a475050b11"
@@ -16,10 +16,9 @@ PrecompileTools = "1"
 julia = "1.6"
 
 [extras]
-AbstractAlgebra = "c3fe647b-3220-5bb0-a1ea-a7954cac585d"
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "Random", "BenchmarkTools", "AbstractAlgebra"]
+test = ["Test", "Random", "BenchmarkTools"]