master:
```
[ Info: Iteration over S8 PermGroup (7 transpositions)
  3.397 ms (120949 allocations: 6.15 MiB)
[ Info: Iteration over K ≅ S8 PermGroup:
  0.000044 seconds (580 allocations: 33.867 KiB)
  3.626 ms (125987 allocations: 6.46 MiB)
  0.000095 seconds (1.21 k allocations: 72.883 KiB)
  6.646 ms (206980 allocations: 11.40 MiB)
[ Info: Testing and benchmarking Schreier-Sims algorithm
[ Info: Rubik cube 2×2×2 group:
  23.003 μs (525 allocations: 34.01 KiB)
  59.552 μs (1231 allocations: 83.65 KiB)
[ Info: Schreier-Sims for Rubik cube 3×3×3 group:
  1.992 ms (26572 allocations: 2.56 MiB)
  8.730 ms (116251 allocations: 11.36 MiB)
[ Info: Schreier-Sims for SL(4,7):
  10.529 ms (28122 allocations: 10.85 MiB)
  74.842 ms (192150 allocations: 77.18 MiB)
[ Info: Schreier-Sims for a direct-product group:
  2.936 ms (9109 allocations: 3.16 MiB)
  79.653 ms (221289 allocations: 77.26 MiB)
[ Info: Iteration over direct-product group of order 192480
  0.163698 seconds (622.62 k allocations: 216.410 MiB, 4.33% gc time)
  0.376034 seconds (1.24 M allocations: 472.942 MiB, 3.87% gc time)
```

this branch
```
[ Info: Iteration over S8 PermGroup (7 transpositions)
3.138 ms (120949 allocations: 6.15 MiB)
[ Info: Iteration over K ≅ S8 PermGroup:
0.000039 seconds (562 allocations: 32.742 KiB)
3.121 ms (120949 allocations: 6.15 MiB)
0.000075 seconds (1.16 k allocations: 70.273 KiB)
5.809 ms (201815 allocations: 11.09 MiB)
[ Info: Testing and benchmarking Schreier-Sims algorithm
[ Info: Rubik cube 2×2×2 group:
18.565 μs (442 allocations: 29.43 KiB)
52.597 μs (1148 allocations: 79.07 KiB)
[ Info: Schreier-Sims for Rubik cube 3×3×3 group:
1.374 ms (24680 allocations: 2.31 MiB)
6.975 ms (111829 allocations: 10.81 MiB)
[ Info: Schreier-Sims for SL(4,7):
4.393 ms (25970 allocations: 10.22 MiB)
53.474 ms (189388 allocations: 76.13 MiB)
[ Info: Schreier-Sims for a direct-product group:
1.228 ms (9104 allocations: 3.15 MiB)
58.843 ms (219502 allocations: 77.17 MiB)
[ Info: Iteration over direct-product group of order 192480
0.061710 seconds (585.73 k allocations: 184.881 MiB, 9.19% gc time)
0.201417 seconds (1.20 M allocations: 441.331 MiB, 6.48% gc time)
```

overall speedup:
```
for (m,b) in zip(master_bench, branch_bench)
    println(Ratio(m,b))
end

(master/branch): 1.08 (btime), 1.00 (nalloc), 1.00 (mem) Iteration over S8 PermGroup (7 transpositions) (Transversal)
(master/branch): 1.16 (btime), 1.04 (nalloc), 1.05 (mem) Iteration over K ≅ S8 PermGroup: (Transversal)
(master/branch): 1.14 (btime), 1.03 (nalloc), 1.03 (mem) Iteration over K ≅ S8 PermGroup: (SchreierTransversal)
(master/branch): 1.24 (btime), 1.19 (nalloc), 1.16 (mem) Rubik cube 2×2×2 group: (Transversal)
(master/branch): 1.13 (btime), 1.07 (nalloc), 1.06 (mem) Rubik cube 2×2×2 group: (SchreierTransversal)
(master/branch): 1.45 (btime), 1.08 (nalloc), 1.11 (mem) Schreier-Sims for Rubik cube 3×3×3 group: (Transversal)
(master/branch): 1.25 (btime), 1.04 (nalloc), 1.05 (mem) Schreier-Sims for Rubik cube 3×3×3 group: (SchreierTransversal)
(master/branch): 2.40 (btime), 1.08 (nalloc), 1.06 (mem) Schreier-Sims for SL(4,7): (Transversal)
(master/branch): 1.40 (btime), 1.01 (nalloc), 1.01 (mem) Schreier-Sims for SL(4,7): (SchreierTransversal)
(master/branch): 2.39 (btime), 1.00 (nalloc), 1.00 (mem) Schreier-Sims for a direct-product group: (Transversal)
(master/branch): 1.35 (btime), 1.01 (nalloc), 1.00 (mem) Schreier-Sims for a direct-product group: (SchreierTransversal)
```
A great success! 🎉